renaming auth.tsx
