+++ src/store/actions/actionTypes.tsx
@@ -5,3 +5,4 @@ export const AUTH_LOGOUT = "AUTH_LOGOUT";
 export const AUTH_ADMIN = "AUTH_ADMIN";
 export const REGISTRATION_SUCCESS = "REGISTRATION_SUCCESS";
 export const REGISTRATION_FAIL = "REGISTRATION FAIL";
+export const SWITCH_CHANGE_PLAN = "SWITCH CHANGE PLAN";
